% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxplots_with_checks}
\alias{boxplots_with_checks}
\title{Boxplots with checks alongside the boxes}
\usage{
boxplots_with_checks(x, traits, factor, checks = NULL, checks_factor = NULL,
  plottype = "Boxplot", boxcolor = "white", stripcolor = "red",
  checks.color = "blue", legend = TRUE, trim.labels = NULL, ...)
}
\arguments{
\item{x}{A data frame containing at least the columns \code{trait} and
\code{factor} and also \code{checks} and \code{checks_factor} if both
specified.}

\item{traits}{Charachter vector with the column names in \code{data} of the
traits to be plotted.}

\item{factor}{Character string with the column name in \code{data} of the
factor by which to plot the trait.}

\item{checks}{(Optional) Charachter vector with the names of the checks in
the \code{checks_factor} column.}

\item{checks_factor}{(Optional) Charachter string with the column name in
\code{data} of the factor that contains the observations to be plotted as
\code{checks}.}

\item{plottype}{(Optional) Character string speciyfing which type of plot to
produce, should be either \code{"Boxplot"} or \code{"Stripplot"}. Default =
\code{"Boxplot"}.}

\item{boxcolor}{(Optional) Character vector specifying the color(s) for the
boxplots. Default = \code{"white"}.}

\item{stripcolor}{(Optional) Character vector specifying the color(s) for
the points in the stripplot. Default = \code{"red"}.}

\item{checks.color}{(Optional) Character string specifying the color(s) for
the checks. Default = \code{"blue"}.}

\item{legend}{(Optional) Logical indicating whether a legend for the
\code{checks} should be added to the plot. Default = \code{TRUE}.}

\item{trim.labels}{(Optional) Numeric value specifying how many characters
of the levels of \code{factor} should be plotted on the X-axis.  If
\code{NULL}, the levels are plotted as they are in \code{x}. Default =
\code{NULL}.}

\item{\dots}{Extra graphical parameters for the plot-function.}
}
\value{
Plot is submitted to the graphical device.
}
\description{
Create boxplots for a sequence of traits for each level of a grouping
factor.
}
\examples{
data(exampleCV)
exampleCV$CHECK <- sample(c(rep(0, 300),
sample(c(1:4), nrow(exampleCV), 20)), 100)
exampleCV$EXPT <- exampleCV$BLOCK <- NA
colnames(exampleCV)[colnames(exampleCV)== "trait"] <- "dummy"
exampleCV$dummy <- abs(exampleCV$dummy)*100
trait <- "dummy"
boxplots_with_checks(x=exampleCV, traits="dummy", factor="LOCATION",
      checks=exampleCV$GERMPLASM[exampleCV$CHECK!=0],
      checks_factor = "GERMPLASM", checks.color = "grey",
     boxcolor="blue", main="Yield by location\\n(all checks)")
}
\author{
Ruud Derijcker, Katrien Baert
}

