% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateH2}
\alias{calculateH2}
\title{Compute the broad-sense heritability from a mixed model analysis with
ASReml.}
\usage{
calculateH2(model, which = c("line-mean", "overall"),
  genetic.factor = "GERMPLASM", by.location = FALSE, data = NULL,
  checks = NULL)
}
\arguments{
\item{model}{An asreml-object, see \code{\link[asreml]{asreml}}.}

\item{which}{A character string/vector specifying which heritability measure
to compute, choose from \code{"line-mean"} and \code{"overall"}.  By default
both measures are computed.}

\item{genetic.factor}{(Optional) Character string specifying the name of the
genetic factor in the model. Default = \code{"GERMPLASM"}.}

\item{by.location}{Logical indicating whether to compute the heritability on
location-basis or across locations. Default = \code{FALSE}.}

\item{data}{(Optional) The data frame used in the preceeding
\code{asreml}-call.}

\item{checks}{(Optional) The names of the check-lines in the
\code{genetic.factor}-column in \code{data}.}
}
\value{
A data frame with the estimated genetic-, GxE- and residual standard
deviations and the computed heritability.
}
\description{
The broad-sense heritability (\eqn{H^2}) is computed from the variance
components from a linear mixed model \describe{
\item{list(list("V_G"))}{the genetic variance components}
\item{list(list("V_E"))}{the residual variance component}
}
}
\details{
The two approaches for calculating \eqn{H^2}: \describe{
\item{which = }{\deqn{H^2 = \frac{V_G}{V_G + V_E}}}
\item{list("\"overall\"")}{ \deqn{H^2 = \frac{V_G}{V_G + V_E}}}
\item{which = }{ \deqn{H^2 = \frac{V_G}{V_G + V_E/r}}}
\item{list("\"overall\"")}{ \deqn{H^2 = \frac{V_G}{V_G + V_E/r}}}
}
}
\author{
Ruud Derijcker, Katrien Baert
}

