% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CVcommit}
\alias{CVcommit}
\title{Function to make a CV scheme based on a commited test set.}
\usage{
CVcommit(trainingSet, validationSet, k)
}
\arguments{
\item{trainingSet}{character vector of the observations in the training set.}

\item{validationSet}{character vector of the observations in the specified test set.}

\item{k}{integer value for the number of folds used in the k-cross-validation.}
}
\value{
named vector of numeric scores showing the assignment of the observations
to their respective set used in the k-fold cross-validation.
}
\description{
Function to make a CV scheme based on a commited test set.
}
\author{
Ruud Derijcker
}
\references{
Based on synbreed's crossVal function
}

