% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CVincompleteTrial}
\alias{CVincompleteTrial}
\title{Function to make a CV scheme based on random sampling of entry IDs
in an incomplete field trial setup.}
\usage{
CVincompleteTrial(ID, factorID, LOCATION, k, exclusive, seed)
}
\arguments{
\item{ID}{character vector of the observation IDs.
The names are a combination of the entry name and the location.}

\item{factorID}{character vector of the entry IDs used in the randomization.}

\item{LOCATION}{character describing the column name representing the location
information.}

\item{k}{integer value for the number of folds used in the k-cross-validation.}

\item{exclusive}{logical whether sampling should be done with replacement. The
argument is passed to the replace argument of the samp.int function as the negation,
i.e. exclusive is TRUE means replace=FALSE, such that the probability of
choosing the next item is proportional to the weights amongst the remaining
items.}

\item{seed}{numeric value for the seed value used for the randomization by the
set.seed function. In this way randomization can be reproduced by the user.
Default is NULL, which uses 123 as value for the seed.}
}
\value{
named vector of numeric scores showing the assignment of the observations
to their respective set used in the k-fold cross-validation.
}
\description{
Function to make a CV scheme based on random sampling of entry IDs
in an incomplete field trial setup.
}
\details{
we developed our own functionality to simulate randomization for an
incomplete field trial, which is based on permutation using the attached
permute function. in the incomplete field trial setup we define an equal set
of entries at every location, where we ensure no overlap in entry IDs
during this process.
}
\examples{
data(exampleCV)
y <- exampleCV[,which(colnames(exampleCV) \%in\% c("GERMPLASM", "LOCATION"))]
colnames(y) <- c("IDUnique","FACTOR")
y$ID <- paste(y$IDUnique, y$FACTOR, sep="_")
y <- na.omit(y)
n <- length(y$ID)
output <- CVincompleteTrial(ID=y$ID, factorID=y$IDUnique, LOCATION=y$FACTOR,
                           seed=123, k=5, exclusive=TRUE)
table(output)
head(output)
}
\author{
Ruud Derijcker
}

