% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-exampleCV}
\alias{data-exampleCV}
\title{Example data for CV functionalities used in thesis}
\usage{
exampleCV
}
\description{
Example data for CV functionalities used in thesis
}
\details{
The example data frame contains 100 entries grown at 5 locations such
that 20 entries are grown per location. The trait was obtained using the
rnorm function. The data frame holds information for the 100 entries, and 5
features represented in the columns and detailed below:
\describe{
     \item{\code{GERMPLASM}:}{The 5 x 20 entry names.}
     \item{\code{LOCATION}:}{Numeric describing which of the 5 locations the
     entry is coming fromm.}
     \item{\code{trait}:}{A value for the trait obtained by rnorm(100, 0, 1).}
     \item{\code{RANGE}:}{The coordinates for the ranges linked to the locations.}
     \item{\code{ROW}:}{The coordinates for the rows linked to the locations.}
}
}
\examples{
data(exampleCV)
head(exampleCV)
}
\author{
Ruud Derijcker
}
\keyword{exampleCV}

