% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-G}
\alias{data-G}
\title{Realization of the G-matrix used in thesis}
\usage{
G
}
\description{
Realization of the G-matrix used in thesis
}
\details{
The realization of the G-matrix was calculated using the numeric
SNP genotype matrix M following the calculation detailed by Van Raden (2008).
The dimension are 296 x 296 where both rows and columns represent the entries.
The elements of the matrix show the genetic relationships between the entries
that are a combination of a row and column coordinate.
}
\examples{
data(G)
str(G)
}
\author{
Ruud Derijcker
}
\seealso{
makeG
}
\keyword{G}

