% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-genotypes}
\alias{data-genotypes}
\title{Raw SNP marker data used in thesis}
\usage{
genotypes
}
\description{
Raw SNP marker data used in thesis
}
\details{
The data comes from a field trial that is part of a cotton
breeding program. The trial was set up in 2012 across 7 locations in the
US Cotton Belt. At every location the same bi-parental BC_3F_2 was grown
together with a number of entries serving as checks. Genotypes were obtained
for the entries on one locations from a set of genome-wide spread SNP
markers which were scored in a bi-allelic way and represented as a combination
of two of the four bases that make up the DNA (e.g. AA, AC, CC). The data frame
holds information about 1088 SNP markers (rows) for 305 entries (columns),
including checks entries.
}
\examples{
data(genotypes)
head(genotypes)
}
\author{
Ruud Derijcker
}
\keyword{genotypes}

