% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-M}
\alias{data-M}
\title{Cleaned numeric SNP marker data used in thesis}
\usage{
M
}
\description{
Cleaned numeric SNP marker data used in thesis
}
\details{
The data comes from a field trial that is part of a cotton
breeding program. The trial was set up in 2012 across 7 locations in the
US Cotton Belt. At every location the same bi-parental BC_3F_2 was grown
together with a number of entries serving as checks. Genotypes were obtained
for the entries on one locations from a set of genome--wide spread SNP
markers which were scored in a bi-allelic way. After quality control,
genotypes for 299 cross-entries were obtained for 1088 SNP markers.
The observed genotype matrix is converted to allelic content by counting
the number of B alleles for every cell in the genotype matrix. When aligning
the phenotype and genotype information with respect to the entries, we reduce
the number of entries to 296 by discarding genotyped entries that have
completely unobserved phenotypes.
}
\examples{
data(M)
str(M)
}
\author{
Ruud Derijcker
}
\keyword{M}

