% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-P}
\alias{data-P}
\title{Modeled cleaned phenotypic data used in thesis}
\usage{
P
}
\description{
Modeled cleaned phenotypic data used in thesis
}
\details{
The data comes from a field trial that is part of a cotton
breeding program. The trial was set up in 2012 across 7 locations in the
US Cotton Belt. At every location the same bi--parental BC_3F_2 was grown
together with a number of entries serving as checks. Yield performance
measurements were averaged per plot. Location 7 was excluded from the
analysis. Next, a linear mixed model was applied to obtain BLUP predictions
for the yield performance measures where the yield performance measures were
adjusted using the information of the checks, and where a spatial (AR1xAR1)
covariance structure was applied for the rows and ranges in the fields. After
aligning the phenotypes with the genotypes, the resulting data frame holds
information of 1768 observations and 19 features. The features are detailed
below and represent the columns in the data frame.
\describe{
     \item{\code{MERGE}:}{Observation names, which are the combined names of
     the entries and the locations.}
     \item{\code{GERMPLASM}:}{The entry names.}
     \item{\code{LOCATION}:}{The name of the locations.}
     \item{\code{OVERALL_RANGE}:}{The range coordinates when all fields are
     seen as being part of one big field, i.e. same reference grid for all fields.}
     \item{\code{OVERALL_ROW}:}{The row coordinates when all fields are
     seen as being part of one big field, i.e. same reference grid for all fields.}
     \item{\code{RANGEROW}:}{A combination of the range and row coordinates.}
     \item{\code{ROW}:}{The coordinates for the rows linked to the locations.
     Here the reference grid is the location itself.}
     \item{\code{RANGE}:}{The coordinates for the ranges linked to the locations.
     Here the reference grid is the location itself.}
     \item{\code{PLOT}:}{The reference to the plot of the observation.}
     \item{\code{YIELD}:}{The average yield performance measures of the plots for
     the respective observations.}
     \item{\code{EXPERIMENT}:}{Name of the experiment.}
     \item{\code{SUBSUBBLOCK}:}{Placeholder in case subsubblocks were defined
     in case of blocking. Not used here.}
     \item{\code{SUBBLOCK}:}{Placeholder in case subblocks were defined
     in case of blocking. Not used here.}
     \item{\code{BLOCK}:}{Placeholder in case blocks were defined in case of
     blocking. Not used here.}
     \item{\code{CHECK}:}{Numeric showing whether the observation is a check.
     Here the checks were already filtered out.}
     \item{\code{ROWf}:}{Factor for ROW.}
     \item{\code{RANGEf}:}{Factor for RANGE.}
     \item{\code{residuals}:}{Residuals obtained after applying the above
     described linear mixed model.}
     \item{\code{predicted.values}:}{BLUP predictions from the fitted above
     described linear mixed model.}
}
}
\examples{
data(P)
str(P)
}
\author{
Ruud Derijcker
}
\keyword{P}

