% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-phenotypes}
\alias{data-phenotypes}
\title{Raw phenotype data used in thesis}
\usage{
phenotypes
}
\description{
Raw phenotype data used in thesis
}
\details{
The data comes from a field trial that is part of a cotton
breeding program. The trial was set up in 2012 across 7 locations in the
US Cotton Belt. At every location the same bi--parental BC_3F_2 was grown
together with a number of entries serving as checks. Yield performance
measurements were obtained per plot. The data frame contains the field design
information and the yield performance values averaged per plot. The data frame
holds information of 2310 observations and 9 features. The features are
detailed below and represent the columns in the data frame.
\describe{
     \item{\code{GERMPLASM}:}{The entry names.}
     \item{\code{LOCATION}:}{The name of the locations.}
     \item{\code{RANGE}:}{The range coordinates when all fields are
     seen as being part of one big field, i.e. same reference grid for all fields.}
     \item{\code{ROW}:}{The row coordinates when all fields are
     seen as being part of one big field, i.e. same reference grid for all fields.}
     \item{\code{RANGEROW}:}{A combination of the range and row coordinates.}
     \item{\code{LOCAL_ROW}:}{The coordinates for the rows linked to the locations.
     Here the reference grid is the location itself.}
     \item{\code{LOCAL_RANGE}:}{The coordinates for the ranges linked to the locations.
     Here the reference grid is the location itself.}
     \item{\code{PLOT}:}{The reference to the plot of the observation.}
     \item{\code{YIELD}:}{The average yield performance measures of the plots for
     the respective observations.}
}
}
\examples{
data(phenotypes)
head(phenotypes)
}
\author{
Ruud Derijcker
}
\keyword{phenotypes}

