% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataframe.to.list}
\alias{dataframe.to.list}
\title{Convert data frame to list}
\usage{
dataframe.to.list(x, by)
}
\arguments{
\item{x}{data frame}

\item{by}{character specifying column name where levels of factor are used to
make the elements of a list}
}
\value{
list with elements named after levels of the column specified by the
by-argument.
}
\description{
Convert data frame to list
}
\author{
Ruud Derijcker
}

