% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.to.dataframe}
\alias{list.to.dataframe}
\title{Transform a list of data frames to one data frame}
\usage{
list.to.dataframe(x, slotname = "Slot", rowname = "Row")
}
\arguments{
\item{x}{A \code{list} with a data frame with equal column names in each
slot and marker names as the row names.}

\item{slotname}{The column name for the original slot names, default =
\code{"Slot"}.}

\item{rowname}{The column name for the original row names, default =
\code{"Row"}.}
}
\value{
A data frame.
}
\description{
Transform a list of data frames to one data frame
}
\author{
Ruud Derijcker
}

