% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LR.test.asreml}
\alias{LR.test.asreml}
\title{Likelihood ratio test for two asreml-models.}
\usage{
LR.test.asreml(model0, model1)
}
\arguments{
\item{model0}{An asreml-object, see \code{\link[asreml]{asreml}}.}

\item{model1}{An asreml-object with the same fixed part as \code{model0} and
a random part that is an extension of the random part of \code{model0}.}
}
\value{
A list with class "\code{htest}" containing the following
components: \describe{ \item{statistic}{The value of the test statistic.}
\item{parameter}{The degrees of freedom for the test statistic.}
\item{p.value}{The p-value for the test.} \item{conf.int}{\code{NULL}}
\item{estimate}{\code{NULL}} \item{null.value}{\code{NULL}}
\item{alternative}{\code{Extended model outperforms the 'null model'.}}
\item{method}{\code{Likelihood ratio test (Chi-square)}}
\item{data.name}{Names of the two models that were tested.} }
}
\description{
Perform a likelihood ratio test for two asreml-models that have an equal
fixed part and a nested random part.
}
\details{
The likelihood ratio test is performed by comparing two times the difference
in log-likelihood with the quantiles if the \eqn{\chi^2}-distribution with
as degrees of freedom the difference in number of variance components
estimated in both models.

If the degrees of freedom equal 1, the obtained p-value is divided by 2.
}
\examples{
library(asreml)
data(oats)
model0 <- asreml(fixed = yield ~ Variety*Nitrogen, random = ~ Blocks:Wplots,
          data = oats)
model1 <- asreml(fixed = yield ~ Variety*Nitrogen, random = ~ Blocks/Wplots,
          data = oats)
LR.test.asreml(model0, model1)
}
\author{
Ruud Derijcker, Katrien Baert
}

