% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_design_rectangular}
\alias{make_design_rectangular}
\title{Add observations to a data frame so that the design is rectangular (per
field)}
\usage{
make_design_rectangular(object, fields, rows = "ROW", ranges = "RANGE")
}
\arguments{
\item{object}{A data frame with at least the columns \code{rows}, \code{ranges}
and \code{fields}, where fields refer to locations.}

\item{fields}{Character string specifying the column name with the fields names
corresponding to the field locations in \code{object}.}

\item{rows}{(Optional) Character string specifying the column name with the
ROW-coordinates in \code{object}.  Default = \code{"ROW"}.}

\item{ranges}{(Optional) Character string specifying the column name with
the RANGE-coordinates in \code{object}.  Default = \code{"RANGE"}.}
}
\value{
An extended version of the data frame \code{object}.
}
\description{
Add observations to a data frame so that the design is rectangular (per
field).
}
\details{
For each level of \code{fields}, a rectangular design is created by adding
empty observations for the missing \code{rows} and \code{ranges}.
}
\examples{
data(exampleCV)
scheme1 <- crossValidate(x=exampleCV, id="GERMPLASM", factor="LOCATION",
                         k=5, replication=3, seed=NULL, exclusive=TRUE,
                         sampling="randomByFactor",verbose=TRUE)
P2 <- data.frame(exampleCV, scheme1)
P2 <- make_design_rectangular(P2, fields="LOCATION",rows="ROW",ranges="RANGE")
head(P2)
}
\author{
Ruud Derijcker
}

