% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeG}
\alias{makeG}
\title{Make genomic relationship matrix using Van Raden (2008)}
\usage{
makeG(x, bAlleleFrequency = NULL)
}
\arguments{
\item{x}{matrix of genotypes in bi-allelic format (e.g. AA, AC, CC). Missing
data is entered as NA.}

\item{bAlleleFrequency}{vector of B allele frequencies for the markers. Default
is NULL which uses the column means of the vector x in the calculation of the
G matrix.}
}
\value{
realization of the G matrix in R matrix format
}
\description{
Make genomic relationship matrix using Van Raden (2008)
}
\details{
The function first calculates the Z matrix using makeZ, then scales
the Z matrix using bAlleleFrequency to obtain the G matrix as defined by
Van Raden (2008).
}
\examples{
data(M)
head(makeG(M))
}
\references{
Van Raden, P.M. (2008). Efficient methods to compute genomic predictions.
Journal of Dairy Science, 91(11): 4412-4423.
}
\seealso{
makeZ
}

