% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeZ}
\alias{makeZ}
\title{Make a centered (scaled) genetic marker matrix}
\usage{
makeZ(M, bAlleleFrequency = NULL, scaled = TRUE, haploid = FALSE,
  SNPinRows = TRUE)
}
\arguments{
\item{M}{matrix of genotypes in bi-allelic format (e.g. AA, AC, CC). Missing
data is entered as NA.}

\item{bAlleleFrequency}{vector of B allele frequencies for the markers. Default
is NULL, which takes the column means of the M matrix or the column means divided
by two in case of haplotypes.}

\item{scaled}{logical whether scaling of genotype score need to be done. Default
is TRUE.}

\item{haploid}{logical wheter marker scores are haplotypes. Default is FALSE.}

\item{SNPinRows}{logical whether markers are in rows. If FALSE markers are
assumed to be present in columns. Default is TRUE.}
}
\value{
realization of the Z matrix in R matrix format
}
\description{
Make a centered (scaled) genetic marker matrix
}
\examples{
data(M)
head(makeZ(M))
}

