% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{MaStatThesisRuud}
\alias{MaStatThesisRuud}
\alias{MaStatThesisRuud-package}
\title{MaStatThesisRuud}
\description{
Introductory comments on MaStatThesisRuud
}
\details{
The package holds the data and user defined functionalities used in
the analyses in the M.Sc. thesis of Ruud Derijcker (see reference 3).

Use the \code{example} function to run examples of the various functions.
To view the code of a function simply type the function name in the console.
}
\author{
Ruud Derijcker <ruud.derijcker@ugent.be>
}
\references{
\describe{
  \item{\code{1}:}{De Los Campos, G., Perez, P.(2014). BGLR: Bayesian
  Generalized Linear Regression. Version 1.0.3.
  (http://CRAN.R-project.org/package=BGLR).}
  \item{\code{2}:} {Jarquin, D. et al. (2014). A reaction norm model for
  genomic selection using high-dimensional genomic and environmental data.
  Theoretical and Applied Genetics 127(3):595-607.}
  \item{\code{3}:} {Derijcker, R. (2015). Investigating incorporation of
  genotype x environment interaction (G x E) for genomic selection in a
  practical setting. Unpublished M.Sc. thesis. University of Ghent: Belgium.}
}
}
\keyword{BGLR}
\keyword{GS}

