% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permute}
\alias{permute}
\title{Function to generate a table of combinations when applying permutation}
\usage{
permute(n, r, v = 1:n)
}
\arguments{
\item{n}{numeric for the set size where to apply the permutations}

\item{r}{numeric for the number of wished combinations for the permutations}

\item{v}{numeric value used to show the number of combinations in the output.
The argument is passed to R's base permute function. We use the default of 1:n,
which is the full set is outputted.}
}
\value{
data frame with C(n,r) x r combinations.
}
\description{
Function to generate a table of combinations when applying permutation
}
\details{
we provide a table of the permutation combinations given by the number
of r-combinations of an n-set, C(n,r).
}
\examples{
head(permute(10,2))
}

