% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCV}
\alias{plotCV}
\title{Plot a CV scheme based on crossValidate function}
\usage{
plotCV(data, scheme, title)
}
\arguments{
\item{data}{a data frame with at least the following information:\describe{
\item{\code{GERMPLASM}:}{Name of then entries.}
\item{\code{LOCATION}:}{Name of the geographic locations of the multi-field
trial.}
}}

\item{scheme}{data frame output from the crossValidate function which was based
on a user decided sampling strategy to use in the cross-validation.}

\item{title}{character defining the title of the plot}
}
\value{
plot based on the ggplot package which shows the splitting of the training
and validation sets for the first replication. Further, the plot is split by the
levels of the chosen factor, which is generally the location.
}
\description{
Function to graphically show a CV scheme made by the crossValidate function
}
\examples{
data(exampleCV)
scheme1 <- crossValidate(x=exampleCV, id="GERMPLASM", factor="LOCATION",
                         k=5, replication=3, seed=NULL, exclusive=TRUE,
                         sampling="randomByFactor",verbose=TRUE)
scheme2 <- crossValidate(x=exampleCV, id="GERMPLASM", factor="LOCATION",
                         k=5, replication=3, seed=NULL, exclusive=TRUE,
                         sampling="incompleteTrial",verbose=TRUE)
scheme3 <- crossValidate(x=exampleCV, id="GERMPLASM", factor="LOCATION",
                         k=5, replication=3, seed=NULL, exclusive=TRUE,
                         sampling="randomAccrossFactor",verbose=TRUE)
scheme4 <- crossValidate(x=exampleCV, id="GERMPLASM", factor="LOCATION",
                         k=5, replication=3, seed=NULL, exclusive=TRUE,
                         sampling="randomWithinFactor",verbose=TRUE)
scheme5 <- crossValidate(x=exampleCV, id="GERMPLASM", factor="LOCATION",
                         k=5, replication=3, seed=NULL, exclusive=TRUE,
                         sampling="randomByID",verbose=TRUE)

plotCV(exampleCV, scheme1, "CV-scheme based on randomByFactor")
plotCV(exampleCV, scheme2, "CV-scheme based on incompleteTrial")
plotCV(exampleCV, scheme3, "CV-scheme based on randomAccrossFactor")
plotCV(exampleCV, scheme4, "CV-scheme based on randomWithinFactor")
plotCV(exampleCV, scheme5, "CV-scheme based on randomByID")
}
\author{
Ruud Derijcker
}

