% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCVThesisData}
\alias{plotCVThesisData}
\title{Plot a CV scheme for the data used in the thesis}
\usage{
plotCVThesisData(data, scheme, title = "CV scheme")
}
\arguments{
\item{data}{a data frame with at least the following information:\describe{
\item{\code{GERMPLASM}:}{Name of then entries.}
\item{\code{LOCATION}:}{Name of the geographic locations of the multi-field
trial.}
\item{\code{ROW}:}{Name of then rows}
\item{\code{RANGE}:}{Name of then ranges}
}}

\item{scheme}{data frame output from the crossValidate function which was based
on a user decided sampling strategy to use in the cross-validation.}

\item{title}{character defining the title of the plot}
}
\value{
plot based on the ggplot package which shows the splitting of the training
and validation sets for the first replication. Further, the plot is split by the
levels of the chosen factor, which is generally the location.
}
\description{
Plot a CV scheme for the data used in the thesis
}
\examples{
data(P)
scheme1 <- crossValidate(x=P, id="GERMPLASM", factor="LOCATION",
                        k=5, replication=3, seed=NULL, exclusive=TRUE,
                        sampling="incompleteTrial",verbose=TRUE)
plotCVThesisData(P, scheme1, "CV scheme 'incompleteTrial' on thesis data set")
}
\author{
Ruud Derijcker
}

