% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotFields}
\alias{plotFields}
\title{Create a field-layout plot from a data frame.}
\usage{
plotFields(x, column, textColumn = NULL, asFactor = FALSE,
  borderColumns = c("EXPT", "BLOCK"), colorscale = NULL, colors = NULL,
  rows = "ROW", ranges = "RANGE", ...)
}
\arguments{
\item{x}{A data frame containing at least the columns \code{column},
\code{rows} and \code{ranges}.}

\item{column}{Character string specifying the name of the column in \code{x}
to be used for coloring the different positions in the field.  This should
be either a factor-column or a numeric column.}

\item{textColumn}{(Optional) Character string specifying the name of the column in
\code{x} to be used to print text within the plots}

\item{asFactor}{Logical indicating whether the specified column should be
considered as a factor, even if it is numeric.}

\item{borderColumns}{(Optional) A chracter vector specifying the factors
(maximum of three factors) to be used to draw border columns in the field map.
Default = \code{c("EXPT", "BLOCK")}.
The factors are assumed to be given in nested order, such as \code{c("EXPT", "TRIAL", "BLOCK")},
\code{c("EXPT", "BLOCK", "SUBBLOCK")} etc.
-- if only one factor is given, a bold border will be drawn
-- if two factors are given, a bold border will be drawn for the main factor with regular border
for the sub-factor
-- if three factors are given, a bold border will be drawn for the main factor, regular border
will be drawn for the sub-factor and dotted-line border will be drawn for sub-sub-factor}

\item{colorscale}{(Optional) A numeric vector of length 2 with the range to
use the \code{colors} over and is applicable only when 'column' is numeric variable.
Default is \code{NULL} in which case the observed range is computed from the \code{column}.}

\item{colors}{Charachter vector speciyfing the colors to be used for the
different levels in \code{column}.  Default = \code{NULL} in which case
pre-specified colors are used.}

\item{rows}{Character string specifying the column name with the
ROW-coordinates in \code{x}.}

\item{ranges}{Character string specifying the column name with the
RANGE-coordinates in \code{x}.}

\item{\dots}{Extra graphical parameters for the
\code{\link[lattice]{levelplot}}-function.}
}
\value{
Plot is submitted to the graphical device.
}
\description{
Create a field-layout plot from a data frame.
}
\examples{
data(exampleCV)
exampleCV$CHECK <- sample(c(rep(0, 80),
sample(c(1:4), nrow(exampleCV), 20)), 100)
exampleCV$EXPT <- exampleCV$BLOCK <- NA
plotFields(x=exampleCV, column="CHECK", main="Fieldplot for dummy example",
           textColumn="GERMPLASM",
           colors=c("grey", "red","blue","green","yellow"))
}
\author{
Ruud Derijcker, Katrien Baert, Bert Lataire
}

