% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot a series of residual plots obtained from a linear mixed model}
\usage{
plotResiduals(x, residuals, by, rows = "ROW", ranges = "RANGE",
  trimVariogram = FALSE, trait = NULL)
}
\arguments{
\item{x}{A data frame that contains at least the columns \code{residuals},
\code{ROW}, \code{RANGE} and \code{by}.}

\item{residuals}{Character string specifying the column name of the
residuals in \code{x}.}

\item{by}{Character string specifying the column name for the splitting
factor in \code{x}, usually this is a location factor.}

\item{rows}{Character string specifying the column name with the \code{ROW}-
coordinates in x. Default = "ROW".}

\item{ranges}{Character string specifying the column name with the
\code{RANGE}-coordinates in x. Default = "RANGE".}

\item{trimVariogram}{Logical indicating whether to trim the variogram to
shorter lag values. Default = "FALSE".}

\item{trait}{Character string specifying the trait name to use in the title for
the plots. Default to \code{NULL}, in which case no secondary title with trait name
is printed.}
}
\value{
Produces various diagnostic residual plots. The colors are scaled to be
comparable over different \code{by}-levels (if there are). Four plots are
outputted per location and are shown in the graphical device. The plots are
listed below and detailed about the construction and use is given in
Appendix A in the referred thesis. \describe{
 \item{plot upper-left}{Variogram of the residuals for the rows and ranges.
 By default the lags are limited by 30. }
 \item{plot upper-right}{Conditional residual plots, colored by their value and
 sign.}
 \item{plot bottom-left}{Scatterplot of the the residuals against the range
  number conditional on the row number}
 \item{plot bottom-right}{Scatterplot of the the residuals against the row
  number conditional on the range number}
}
}
\description{
Plot a series of residual plots obtained from a linear mixed model
}
\author{
Ruud Derijcker, Krishna Bondalapati
}
\references{
Derijcker, R. (2015). Investigating incorporation of
genotype x environment interaction (G x E) for genomic selection in a
practical setting. Unpublished M.Sc. thesis. University of Ghent: Belgium.
}
\seealso{
\code{\link{plotVariogram}}.
}

