% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotVariogram}
\alias{plotVariogram}
\title{Plot variogram from residuals}
\usage{
plotVariogram(x, trim = FALSE, title = "Variogram")
}
\arguments{
\item{x}{A data frame that contains at least the columns \code{residuals},
\code{ROW} and \code{RANGE}.}

\item{trim}{Logical indicating whether to trim the variogram to shorter lag
values or not. Default = \code{FALSE}.}

\item{title}{Character string specifying the title for the plot.
Default = \code{"Variogram"}.}
}
\value{
Plot is submitted to the graphical device.
}
\description{
Plot variogram from residuals
}
\author{
Ruud Derijcker, Krishna Bondalapati
}

