% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizeInference}
\alias{summarizeInference}
\title{Summarize the inference results from  from inferenceBGLR}
\usage{
summarizeInference(P, obj)
}
\arguments{
\item{P}{a data frame that holds the design information and phenotypes for the
data to be modeled. The data should hold following features, which are
detailed below and represent the columns in the data frame. For inference on
the thesis data set we use the data set P obtained by typing data(P) in the
console.}

\item{obj}{object obtained from the inferenceBGLR function.}
}
\value{
A list of summarized inference measures from \code{obj}.
}
\description{
Summarize the inference results from  from inferenceBGLR
}
\details{
The function summarizes the obtained results for the bias, MSPE, and
predictive ability ("correlation") using the average across the different folds
and replications as  was specified in the cross-validation and \code{inferenceBGLR}
function. Confidence intervals for the predictive ability were calculated using
rho +/- 1.96*sqrt((1-rho^2)/(nrow(P)-2)). The confidence intervals for the bias
and MSPE were calculated using a minimum and maximum value calculated as
mean(metric) +/- SE(metric), following \code{synbreed} reporting by T. Albrecht
in the \code{summary.cvData} function. See also reference.
}
\examples{
data(G)
data(P)
scheme <- crossValidate(x=P, id="GERMPLASM", factor="LOCATION", k=5,
                        replication=2, seed=NULL, exclusive=TRUE,
                        sampling="randomByID",verbose=TRUE)
output <- inferenceBGLR(P, CVscheme=scheme, modelName="modelG", id="GERMPLASM",
                       G=G, factor="LOCATION", trait="YIELD", nIter=1500, burnIn=250,
                       replications=2)
output <- summarizeInference(P, output)
output
}
\author{
Ruud Derijcker
}
\references{
Wimmer, V., et al. (2012) synbreed: a framework for the analysis
of genomic prediction data using R. Bioinformatics, 28: 2086-2087.
}

